+++ src/crypto/randomx/vm_interpreted_light.cpp
@@ -39,7 +39,7 @@ namespace randomx {
 
 	template<int softAes>
 	void InterpretedLightVm<softAes>::datasetRead(uint64_t address, int_reg_t(&r)[8]) {
-		uint32_t itemNumber = address / CacheLineSize;
+		uint32_t itemNumber = static_cast<uint32_t>(address / CacheLineSize);
 		int_reg_t rl[8];
 		
 		initDatasetItem(cachePtr, (uint8_t*)rl, itemNumber);
